+++ src/anemoi/transform/filters/repeat_members.py
@@ -10,7 +10,6 @@
 
 import logging
 from typing import Any
-from typing import List
 
 import earthkit.data as ekd
 
@@ -22,7 +21,7 @@ from anemoi.transform.filters import filter_registry
 LOG = logging.getLogger(__name__)
 
 
-def make_list_int(value: Any) -> List[int]:
+def make_list_int(value: Any) -> list[int]:
     """Convert a value to a list of integers.
 
     Parameters
@@ -37,7 +36,7 @@ def make_list_int(value: Any) -> List[int]:
     """
     if isinstance(value, str):
         if "/" not in value:
-            return [value]
+            return [int(value)]
         bits = value.split("/")
         if len(bits) == 3 and bits[1].lower() == "to":
             value = list(range(int(bits[0]), int(bits[2]) + 1, 1))
@@ -48,7 +47,7 @@ def make_list_int(value: Any) -> List[int]:
     if isinstance(value, list):
         return value
     if isinstance(value, tuple):
-        return value
+        return list(value)
     if isinstance(value, int):
         return [value]
 
@@ -68,9 +67,9 @@ class RepeatMembers(Filter):
     def __init__(
         self,
         *,
-        numbers: List[int] = None,
-        members: List[int] = None,
-        count: int = None,
+        numbers: list[int] | None = None,
+        members: list[int] | None = None,
+        count: int | None = None,
     ) -> None:
         """Parameters
         -------------