+++ src/anemoi/transform/grouping/__init__.py
@@ -8,13 +8,15 @@
 # nor does it submit to any jurisdiction.
 
 
+import logging
 from collections import defaultdict
+from collections.abc import Callable
+from collections.abc import Iterator
 from typing import Any
-from typing import Callable
-from typing import Dict
-from typing import Iterator
-from typing import List
-from typing import Tuple
+
+from earthkit.data import SimpleFieldList
+
+LOG = logging.getLogger(__name__)
 
 
 def _lost(f: Any) -> None:
@@ -28,13 +30,13 @@ def _lost(f: Any) -> None:
     raise ValueError(f"Lost field {f}")
 
 
-def _flatten(params: List[Any]) -> List[str]:
+def _flatten(params: list[Any] | tuple[Any, ...]) -> list[str]:
     """Flatten a list of parameters.
 
     Parameters
     ----------
-    params : list of Any
-        List of parameters to flatten.
+    params : list or tuple of Any
+        List or tuple of parameters to flatten.
 
     Returns
     -------
@@ -59,12 +61,38 @@ class GroupByParam:
         List of parameters to group by.
     """
 
-    def __init__(self, params: List[str]) -> None:
+    def __init__(self, params: list[str]) -> None:
         if not isinstance(params, (list, tuple)):
             params = [params]
         self.params = _flatten(params)
 
-    def iterate(self, data: List[Any], *, other: Callable[[Any], None] = _lost) -> Iterator[Tuple[Any, ...]]:
+    def _get_groups(self, data: list[Any], *, other: Callable[[Any], None] = _lost) -> None:
+        assert callable(other), type(other)
+        self.groups: dict[tuple[tuple[str, Any], ...], dict[str, Any]] = defaultdict(dict)
+        self.groups_params = set()
+        for f in data:
+            key = f.metadata(namespace="mars")
+            if not key:
+                keys = [k for k in f.metadata().keys() if k not in ("latitudes", "longitudes", "values")]
+                key = {k: f.metadata(k) for k in keys}
+                if not keys:
+                    raise NotImplementedError(f"GroupByParam: {f} has no sufficient metadata")
+
+            param = key.pop("param", f.metadata("param"))
+
+            if param not in self.params:
+                other(f)
+                continue
+
+            key = tuple(key.items())
+
+            if param in self.groups[key]:
+                raise ValueError(f"Duplicate component {param} for {key}")
+            self.groups[key][param] = f
+            self.groups_params.add(param)
+        LOG.info(f"Params groups: {self.groups_params}")
+
+    def iterate(self, data: list[Any], *, other: Callable[[Any], None] = _lost) -> Iterator[tuple[Any, ...]]:
         """Iterate over the data and group fields by parameters.
 
         Parameters
@@ -79,28 +107,54 @@ class GroupByParam:
         Iterator[Tuple[Any, ...]]
             Iterator yielding tuples of grouped fields.
         """
-        assert callable(other), type(other)
-        groups: Dict[Tuple[Tuple[str, Any], ...], Dict[str, Any]] = defaultdict(dict)
+        self._get_groups(data, other=other)
+        for _, group in self.groups.items():
+            if len(group) != len(self.params):
+                for p in data:
+                    print(p)
+                raise ValueError(f"Missing component. Want {sorted(self.params)}, got {sorted(self.groups.keys())}")
 
+            yield tuple(group[p] for p in self.params)
+
+
+class GroupByParamVertical(GroupByParam):
+    def _get_groups(self, data: list[Any], *, other: Callable[[Any], None] = _lost) -> None:
+        assert callable(other), type(other)
+        self.groups: dict[tuple[tuple[str, Any], ...], dict[str, Any]] = defaultdict(dict)
+        self.groups_params = set()
+        levels: dict[str, Any] = defaultdict(list)
         for f in data:
             key = f.metadata(namespace="mars")
+            if not key:
+                keys = [k for k in f.metadata().keys() if k not in ("latitudes", "longitudes", "values")]
+                key = {k: f.metadata(k) for k in keys}
+                if not keys:
+                    raise NotImplementedError(f"GroupByParam: {f} has no sufficient metadata")
+
             param = key.pop("param", f.metadata("param"))
+            _ = key.pop("levtype", None)
+            level = key.pop("levelist", None)
 
             if param not in self.params:
                 other(f)
                 continue
 
-            key = tuple(key.items())
+            key = tuple(sorted(tuple(key.items())))
 
-            if param in groups[key]:
-                raise ValueError(f"Duplicate component {param} for {key}")
-
-            groups[key][param] = f
-
-        for _, group in groups.items():
-            if len(group) != len(self.params):
-                for p in data:
-                    print(p)
-                raise ValueError(f"Missing component. Want {sorted(self.params)}, got {sorted(group.keys())}")
-
-            yield tuple(group[p] for p in self.params)
+            if level is None:
+                if param in self.groups[key]:
+                    raise ValueError(f"Duplicate component {param} for {key}")
+                self.groups[key][param] = f
+            else:
+                if param in self.groups[key]:
+                    if level in levels[param]:
+                        raise ValueError(f"Duplicate component {param} for {key} and level {level}")
+                    else:
+                        self.groups[key][param].append(f)
+                else:
+                    ds = SimpleFieldList()
+                    ds.append(f)
+                    self.groups[key][param] = ds
+                levels[param].append(level)
+            self.groups_params.add(param)
+        LOG.info(f"Params groups: {self.groups_params}")